beneighb/requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
black==23.3.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
cryptography==41.0.3
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.2
django-allauth==0.54.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
factory-boy==3.3.0
Faker==19.2.0
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
ipdb==0.13.13
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.8.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.7.0
urllib3==2.0.4
wcwidth==0.2.6
django-extensions==3.2.3
google-api-python-client==2.99.0

# Redis
redis==5.0.1
redis-cli==1.0.1
celery==5.3.6
firebase-admin==6.4.0
